CLEAN: add back getPointAtTargetAngle and clean up
